refactor: make useFriendModal hook
모달에서 하는일이 많아서 hook으로 빼줌
