+++ front/src/components/FriendsSection/FriendList/FriendModal/FriendModal.js
@@ -1,12 +1,9 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import makeModal from 'components/globalComponents/Modal/Modal';
 import Button from 'components/globalComponents/Button/Button';
 import { ButtonSection } from 'components/FriendsSection/FriendList/FriendModal/FriendModal.style';
 import PropTypes from 'prop-types';
-import { DELETE_FRIEND } from 'queries/friend';
-import { useMutation } from '@apollo/react-hooks';
-import { SEND_FRIEND_REQUEST } from 'queries/beforeFriend';
-import GlobalContext from 'global.context';
+import useFriendModal from 'hooks/FriendsSection/useFriendModal';
 
 FriendModal.propTypes = {
   modalContent: PropTypes.shape({
@@ -30,66 +27,11 @@ export default function FriendModal({
   dispatchModalContent,
   refetch,
 }) {
-  const { onlineSocket, user } = useContext(GlobalContext);
-  const [deleteFriendRequest] = useMutation(DELETE_FRIEND, {
-    onCompleted({ deleteFriend }) {
-      onlineSocket.emit('deleteFriend', deleteFriend);
-      dispatchModalContent({
-        type: 'deleteDone',
-        nickname: modalContent.nickname,
-      });
-      refetch();
-    },
-    onError() {
-      dispatchModalContent({
-        type: 'error',
-        content: '에러가 발생했습니다.',
-      });
-    },
+  const [clearModalContent, clickOKButton] = useFriendModal({
+    modalContent,
+    dispatchModalContent,
+    refetch,
   });
-
-  const [sendFriendRequest] = useMutation(SEND_FRIEND_REQUEST, {
-    onCompleted({ sendFriendRequest: friend }) {
-      onlineSocket.emit('alarm', {
-        user: friend,
-        message: `${user.nickname}님이 친구가 되고 싶어해요`,
-      });
-      dispatchModalContent({
-        type: 'addDone',
-        nickname: friend.nickname,
-      });
-    },
-    onError({ graphQLErrors }) {
-      dispatchModalContent({
-        type: 'error',
-        content: graphQLErrors[0].message,
-      });
-    },
-  });
-
-  const clearModalContent = () => {
-    dispatchModalContent({ type: 'clear' });
-  };
-
-  const clickOKButton = () => {
-    switch (modalContent.current) {
-      case 'addDone':
-      case 'deleteDone':
-      case 'error':
-        return dispatchModalContent({ type: 'clear' });
-      case 'addRequest':
-        return sendFriendRequest({
-          variables: { nickname: modalContent.nickname },
-        });
-      case 'deleteRequest':
-        return deleteFriendRequest({
-          variables: { id: modalContent.id },
-        });
-      default:
-        throw new Error(`${modalContent.current}cannot find current`);
-    }
-  };
-
   const Body = () => <span>{modalContent.content}</span>;
   const Footer = () => (
     <ButtonSection>
